Bug-fix for engine_from_config()
The function was storing its parsed values in the same dict-like object,
that it received as paremeter. This was not only modifying an external
resource, but was also leading to an error, if the passed type was a
ConfigParser object, which contained a 'poolclass' definition for
example. The object would complain, stating that it only accepts string
values.